[WIP] extending sandboxed function wrapper
I think we can now just override a pointer in memory of the function binary and make it point to Dandelion's Context for the function invocation. It then seems convenient to implement moving the data and constructing the runtime data in wasm memory inside the _start() function of the wrapper.