copy system data and duplicate SDK's system data struct
While there is a pre-allocated system data region in wasm memory, the function SDK's __system_data struct is allocated somewhere else in wasm memory. This implementation expects a bytes array in the wasm entry point holding all system data (including the top level __system_data struct at the beginning) and initializes the system data in the wasm module as follows

1. it copies tje array as a whole to the pre-allocated region in wasm memory
2. it clones the top level struct and puts it where __system_data is located in wasm memory
3. it fixes the pointers in the newly cloned __system_data struct by adding the offset between that location and the system data region

It seems like I don't need all of this, committing it now to have it in source control.