wasm sdk heap implementation
I also changed the function entry point to receive the system DandelionSystemData struct directly.

This implementation places the function SDK's virtual heap at the end of the wasm module's heap, by extending the latter in rWasm beyond the minimum.

While conceptually working, I missed that the SDK actually uses dandelion_alloc() to place outputs, which will end up in the SDK's heap and not in the pre-allocated system data region. I might therefore move the system data region into (or right before) the SDK heap to have only one consecutive chunk of memory in Dandelion to interface with the function.