external memory support
The entry point now receives a reference to memory that will be used to store the wasm memory. This memory is still fixed in size, but can reside on heap as well.

I also made the sdk mem size (and also full wasm mem size) environment variables which the sandboxed-function crate will parse automatically.